Scrap the idea of AutoIPython (at least for now).
IPython's interactive input code changes drastically from 4.x to 5.x
and I can't figure out how to control Python Prompt Toolkit.

Also, pagination or the new multiline strings handling borked
the presentation code used for 4.x.

I'm not even sure that the idea of AutoIPython is even sound so
I decided to scrap it until I find a way to make it work.

I have left in place the infrastructure for multiple backends just
in case I need to add a new backend in the future.